# packages/backend/openapi3.yaml
openapi: 3.0.1
info:
  title: detiler
  description: api for managing tiles' details
  version: 1.0.0
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
tags:
  - name: detail
    description: detail selected tile
  - name: kit
    description: existing kits
security: []
paths:
  /kits:
    get:
      operationId: getKits
      tags:
        - kit
      summary: Get a list of available kits
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
        5XX:
          $ref: '#/components/responses/UnexpectedError'
    post:
      summary: Register a new kit
      requestBody:
        description: kit to be created
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/kit'
      operationId: createKit
      tags:
        - kit
      responses:
        '201':
          description: created
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
        5XX:
          $ref: '#/components/responses/UnexpectedError'
  /detail:
    get:
      operationId: getTileDetails
      tags:
        - detail
      summary: return tiles' details by query
      parameters:
        - in: query
          name: kits
          description: return tile details in specified kits
          schema:
            type: array
            items:
              $ref: '#/components/schemas/kitName'
            minItems: 1
            default: []
            uniqueItems: true
        - in: query
          name: minZoom
          description: min zoom level
          schema:
            $ref: '#/components/schemas/z'
          required: true
        - in: query
          name: maxZoom
          description: max zoom level
          schema:
            $ref: '#/components/schemas/z'
          required: true
        - in: query
          name: bbox
          description: tiles within bounding box
          schema:
            type: array
            items:
              anyOf:
                - $ref: '#/components/schemas/longtitude'
                - $ref: '#/components/schemas/latitude'
            minItems: 4
            maxItems: 4
          required: true
        - in: query
          name: from
          description: pagination query from
          schema:
            $ref: '#/components/schemas/from'
        - in: query
          name: size
          description: pagination query size limitation
          schema:
            $ref: '#/components/schemas/size'
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/tileDetails'
        400:
          $ref: '#/components/responses/BadRequest'
        5XX:
          $ref: '#/components/responses/UnexpectedError'
  /detail/{z}/{x}/{y}:
    get:
      operationId: getTileDetailsByKits
      tags:
        - detail
      summary: return tiles' details in selected kits
      parameters:
        - in: path
          name: z
          description: zoom level
          schema:
            $ref: '#/components/schemas/z'
          required: true
        - in: path
          name: x
          description: x value
          schema:
            $ref: '#/components/schemas/x'
          required: true
        - in: path
          name: y
          description: y value
          schema:
            $ref: '#/components/schemas/y'
          required: true
        - in: query
          name: kits
          description: return tile details in specified kits
          schema:
            type: array
            items:
              $ref: '#/components/schemas/kitName'
            uniqueItems: true
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/tileDetails'
        400:
          $ref: '#/components/responses/BadRequest'
        5XX:
          $ref: '#/components/responses/UnexpectedError'
  /detail/{kit}/{z}/{x}/{y}:
    get:
      operationId: getTileDetailsByKit
      tags:
        - detail
      summary: return tile details in a single kit
      parameters:
        - in: path
          name: kit
          description: kit name
          schema:
            $ref: '#/components/schemas/kitName'
          required: true
        - in: path
          name: z
          description: zoom level
          schema:
            $ref: '#/components/schemas/z'
          required: true
        - in: path
          name: x
          description: x value
          schema:
            $ref: '#/components/schemas/x'
          required: true
        - in: path
          name: y
          description: y value
          schema:
            $ref: '#/components/schemas/y'
          required: true
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/tileDetails'
        400:
          $ref: '#/components/responses/BadRequest'
        404:
          $ref: '#/components/responses/NotFound'
        5XX:
          $ref: '#/components/responses/UnexpectedError'
    put:
      operationId: putTileDetails
      tags:
        - detail
      summary: upsert a given tile details
      parameters:
        - in: path
          name: kit
          description: kit name
          schema:
            $ref: '#/components/schemas/kitName'
          required: true
        - in: path
          name: z
          description: zoom level
          schema:
            $ref: '#/components/schemas/z'
          required: true
        - in: path
          name: x
          description: x value
          schema:
            $ref: '#/components/schemas/x'
          required: true
        - in: path
          name: y
          description: y value
          schema:
            $ref: '#/components/schemas/y'
          required: true
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/updatableTileDetails'
      responses:
        201:
          description: Created
        204:
          description: Updated
        404:
          $ref: '#/components/responses/NotFound'
        400:
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
components:
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/error'
    NotFound:
      description: Not Found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/error'
    Conflict:
      description: Conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/error'
    UnexpectedError:
      description: Unexpected Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/error'
  schemas:
    error:
      type: object
      required:
        - message
      properties:
        message:
          type: string
    longtitude:
      type: number
      minimum: -180
      maximum: 180
      exclusiveMinimum: true
    latitude:
      type: number
      minimum: -85.05112877980659
      maximum: 85.05112877980659
    from:
      type: number
      minimum: 0
      default: 0
    size:
      type: number
      minimum: 0
      maximum: 1000
      default: 100
    z:
      type: integer
      minimum: 0
    x:
      type: integer
      minimum: 0
    y:
      type: integer
      minimum: 0
    kitName:
      type: string
      minLength: 1
    state:
      type: integer
      minimum: -1
    unixTimestamp:
      type: integer
      minimum: 0
    counter:
      type: integer
      minimum: 0
    coordinates:
      type: string
      pattern: '^(\d+(\.\d+)?),(\d+(\.\d+)?)$'
      example: '34.114,35.123'
    updatableTileDetails:
      type: object
      properties:
        timestamp:
          $ref: '#/components/schemas/unixTimestamp'
        state:
          $ref: '#/components/schemas/state'
      required:
        - timestamp
      additionalProperties: false
    kit:
      type: object
      properties:
        name:
          $ref: '#/components/schemas/kitName'
      required:
        - name
    tileDetails:
      type: object
      properties:
        z:
          $ref: '#/components/schemas/z'
        x:
          $ref: '#/components/schemas/x'
        y:
          $ref: '#/components/schemas/y'
        kit:
          $ref: '#/components/schemas/kitName'
        state:
          $ref: '#/components/schemas/state'
        createdAt:
          $ref: '#/components/schemas/unixTimestamp'
        updatedAt:
          $ref: '#/components/schemas/unixTimestamp'
        updateCount:
          $ref: '#/components/schemas/counter'
        coordinates:
          $ref: '#/components/schemas/coordinates'
      required:
        - z
        - x
        - y
        - kit
        - state
        - createdAt
        - updatedAt
        - updateCount
        - coordinates